requirements.txt
pandas
# bs4
# requests
python-dotenv
matplotlib
# scikit-learn
ipykernel
jupyterlab